apps/core-api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.5
pydantic-settings==2.7.1
email-validator==2.2.0
python-json-logger==3.2.1
python-jose==3.3.0
passlib==1.7.4
httpx==0.28.1
celery==5.4.0
redis==5.2.1
PyYAML==6.0.2
jinja2==3.1.5
jsonschema==4.23.0
elasticsearch==8.17.2
